@@ -27,8 +27,9 @@ from typing import (
 from libcst import CSTNode, CSTNodeT, FlattenSentinel, RemovalSentinel
 from libcst._add_slots import add_slots
 from libcst.metadata import CodePosition as CodePosition, CodeRange as CodeRange
-
 from packaging.version import Version
+
+__all__ = ("Version",)
 
 T = TypeVar("T")
 
@@ -48,9 +49,7 @@ Timings = Dict[str, int]
 TimingsHook = Callable[[Timings], None]
 
 VisitorMethod = Callable[[CSTNode], None]
-VisitHook = Callable[[str], ContextManager]
-
-Version
+VisitHook = Callable[[str], ContextManager[None]]
 
 
 @dataclass(frozen=True)
@@ -212,7 +211,7 @@ class Config:
     # post-run processing
     formatter: Optional[str] = None
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         self.path = self.path.resolve()
         self.root = self.root.resolve()
 
@@ -222,7 +221,7 @@ class RawConfig:
     path: Path
     data: Dict[str, Any]
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         self.path = self.path.resolve()
 
 
@@ -237,7 +236,7 @@ class LintViolation:
     range: CodeRange
     message: str
     node: CSTNode
-    replacement: Optional[NodeReplacement]
+    replacement: Optional[NodeReplacement[CSTNode]]
     diff: str = ""
 
     @property
