@@ -3,8 +3,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from fixit.config import collect_rules, Config
-from fixit.ftypes import QualifiedRule
+from fixit.config import collect_rules
+from fixit.ftypes import Config, QualifiedRule
 
 from fixit.testing import add_lint_rule_tests_to_module
 from .config import ConfigTest
